+++ src/core/utils/index.js
@@ -5,6 +5,7 @@ import copyText from './copy-text.js';
 import debounce from './debounce.js';
 import * as dom from './dom.js';
 import * as html from './html.js';
+import injectStyles from './inject-styles.js';
 import * as json from './json.js';
 import * as layout from './layout.js';
 import * as dataLoad from './load-data.js';
@@ -24,6 +25,7 @@ export default {
     debounce,
     ...dom,
     ...html,
+    injectStyles,
     ...json,
     ...layout,
     pattern,