+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Added "Views showcase" to header of view editor form on report page
+
 ## 1.0.0-beta.54 (15-02-2021)
 
 - Fixed premature render when `Widget#setDataProgress()` is used, which resulted in double page rendering