+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Added "Views showcase" to header of view editor form on report page
+- Exposed `injectStyles()` helper
 
 ## 1.0.0-beta.54 (15-02-2021)